Added Files and some documents about the game.
I will delete the source of this game if Konami does request it. This is for education purpose. This might help do study and understand the game so you can build a patches or make hacks. You can build the original game and you are supposed to own the game to use this.